On a cluster node with Spack-installed Gtest (version 1.12.1) I have issues...

See merge request zrc-von-neumann-lab/runtime-system-innovations/hicr!119

+++ tests/backends/host/hwloc/L1/memoryManager.cpp
@@ -53,13 +53,13 @@ TEST(MemoryManager, Memory)
   // Getting first device found
   HiCR::L0::Topology::deviceList_t dList;
   EXPECT_NO_THROW(dList = t.getDevices());
-  ASSERT_GT(dList.size(), 0);
+  ASSERT_GT(dList.size(), (unsigned int)0);
   auto d = *dList.begin();
 
   // Getting memory resource list (should be size 1)
   HiCR::L0::Device::memorySpaceList_t mList;
   EXPECT_NO_THROW(mList = d->getMemorySpaceList());
-  ASSERT_GT(mList.size(), 0);
+  ASSERT_GT(mList.size(), (unsigned int)0);
 
   // Getting memory resource
   auto r = *mList.begin();
